requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
cryptography==42.0.5
cutie==0.3.2
Deprecated==1.2.14
discord.py==2.3.2
frozenlist==1.4.1
gcovr==7.2
idna==3.6
intervaltree==3.1.0
Jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
multidict==6.0.5
pillow==10.2.0
pycparser==2.21
PyGithub==1.58.2
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pytz==2024.1
readchar==4.0.5
requests==2.31.0
sortedcontainers==2.4.0
urllib3==2.2.1
watchdog==4.0.0
windows-curses==2.3.2; platform_system == "Windows"
wrapt==1.16.0
yarl==1.9.4
